Update ASFF parser
My goal is to make the ASFF findings more actionable than what is currently being added into DefectDojo fields.  In the other SecurityHub parser it was enhanced to add ResourceId into the description.  I want to also do this with the ASFF findings, in a similar way.

Also, in my opinion (and this change can be removed no worries) but -  vuln_id_from_tool is more appropriate for matching to the security hub finding Id/Arn than the 'unique_id_from_tool'.

unique_id_from_tool shouldn't match a vuln_id (which is the finding_id from AWS) - but instead maybe could be matched to a ResourceId.  Ideally this would be a host or an endpoint, but not always will a Finding have a host associated with it.  HOWEVER, a finding will >always< have a ResourceId and ARN from AWS.  So other than adding into the Description, I'm not sure which other field makes sense to add this to.  I can remove the unique_id_from_tool and vuln_id_from_tool changes if those are not ideal.